refactor: move PortfolioDetail.jsx to pages/, api/, and store/ directories
