+++ frontend/src/api/getPortfolioCans.test.js
@@ -1,7 +1,7 @@
 import { getPortfolioCansAndSetState } from "./getPortfolioCans";
-import store from "../../../store";
-import TestApplicationContext from "../../../applicationContext/TestApplicationContext";
-import { dispatchUsecase } from "../../../helpers/test";
+import store from "../store";
+import TestApplicationContext from "../applicationContext/TestApplicationContext";
+import { dispatchUsecase } from "../helpers/test";
 
 test("successfully gets the Portfolio CAN from the backend and directly puts it into state", async () => {
     const mockCanId = "G99IA14";
