+++ frontend/src/store/portfolioDetailSlice.js
